cleaned up message reactions
- moved reaction code to 'utils/getReply.js'
- adjusted index.js accordingly
