+++ index.js
@@ -2,10 +2,11 @@
 
 const fs = require('node:fs');
 const path = require('node:path');
-const { Client, Collection, Events, GatewayIntentBits, TextChannel } = require('discord.js');
+const { Client, Collection, Events, GatewayIntentBits } = require('discord.js');
 const { useTestBot } = require('./tokenSelection.json');
 const config = require('./config.json');
 const {dirkQuote} = require('./utils/bibleQuotes.js');
+const {getReply} = require('./utils/getReply.js')
 
 const client = new Client({ intents: [GatewayIntentBits.Guilds, GatewayIntentBits.GuildMessages, GatewayIntentBits.MessageContent]});
 
@@ -39,37 +40,8 @@ client.on(Events.InteractionCreate, async interaction => {
 });
 
 client.on('messageCreate', async (message) =>{
-    if(message.author.bot) return;
-    if(message.content.toLowerCase().includes("dirk")){
-        message.reply(dirkQuote());
-    }
-
-	if(message.content.toLowerCase().includes("rolf")||message.content.toLowerCase().includes("betz")){
-        message.reply("Diese Information finden Sie im FORUM!\n (https://ilias.h-ka.de/ilias.php?ref_id=726140&cmd=showThreads&cmdClass=ilrepositorygui&cmdNode=xm&baseClass=ilrepositorygui)");
-    }
-
-	if(message.content.toLowerCase().includes("rolle")||message.content.toLowerCase().includes("wolle")){
-		message.reply({
-			files: [{
-    	attachment: 'gods work/AROUSINGANDEDUCATIONAL2.png',
-    	name: 'AROUSINGANDEDUCATIONAL2.png',
-    	description: 'NOT SAFE FOR WORK'
-			}]
-		});
-	}
-	if(message.content.toLowerCase().includes("pasta")){
-        var pastaDB = fs.readFileSync(`resources/PastaDB.txt`).toString().split("\n");
-		    message.reply(pastaDB[Math.floor(Math.random()*pastaDB.length)])
-    }
-	if(message.content.toLowerCase().includes("assani")||message.content.toLowerCase().includes("mathe")||message.content.toLowerCase().includes("sicher")){
-		message.reply({
-			files: [{
-    	attachment: 'gods work/sicher.png',
-    	name: 'sicher.png',
-    	description: 'SICHER'
-			}]
-		});
-	}
+	if(message.author.bot) return;
+	getReply(message);
 });
 
 client.login(useTestBot ? config.tokenTest : config.token);
